social/pyproject.toml
[tool.poetry]
name = "social_network"
version = "0.1.0"
description = ""
authors = ["Grigoriy <grigoriy.program@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
instaloader = "^4.9"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
fastapi = "^0.75.2"
python-dotenv = "^0.20.0"
cachetools = "^5.0.0"
pyhumps = "^3.5.3"
beautifulsoup4 = "^4.11.1"
loguru = "^0.6.0"
gunicorn = "^20.1.0"
pymongo = "^4.1.1"
bson = "^0.5.10"
Flask = "^2.1.1"
playwright = "^1.21.0"
lxml = "^4.8.0"
sqlmodel = "^0.0.6"
SQLAlchemy = "1.4.35"
requests = "^2.27.1"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
Pygments = "^2.11.2"
isort = "^5.10.1"
black = "^22.3.0"
tqdm = "^4.64.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
preview = true
line_length = 88

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
